Added filtering for finding local chromecasts.
If you have more than one chromecast, figuring out which ip address
is the one you want to talk to isn't obvious.

I added a set of filtering functions so that it's a little easier
to find the chromecast that you want to talk to.

E.g.:
get_single_chromecast(friendly_name="Living Room") will return the
chromecast with the name "Living Room" to you.

Filtering is enabled for ip address, anything in the device_status
or app_status tuples in the PyChromecast object.