Increased Windows Compatibility
Windows protobuf descriptors don’t seem to accept the syntax argument.
Removing it seems to cause to issues. This update will remove the
syntax argument if it is not present in the FileDescriptor. It will
assume that if the syntax argument is not in one, then it is not any of
the Descriptors.