Handle socket/SSL timeouts gracefully
If we get a timeout exception when receiving incoming data we need to
continue the receiving loop.
Older SSL implementations does not raise a socket.timeout exception
but rather ssl.SSLError, we then need to inspect the message to
figure out if it is a timeout.

In addition the stop signal is checked in every receive loop to ensure
we can break out as soon as possible.