Added support for manually stopping the chromecast connection
Calling cast.disconnect() will cause the socket connection to stop
as soon as possible (socket read or message parsing).
It will also wait for the client to finish and join with the thread
unless blocking is False.

Also added a 'join' method to allow the caller to wait for the cast
connection to finish.

The stop detection is improved by checking the 'stop' Event after
parsing and handling the message and also after reading from the socket.
This ensures that the processing loop stops as soon as possible, but
if a message has been fully received we still allow it to be sent to
any handlers.