Check the stop flag in the reconnection loop
If the stop flag has been set it needs to break out of the reconnect
loop as soon as possible. Without this it would try to reconnect
forever even if it was told to stop.

Also handle ChromecastConnectionError in the socket client main loop
to ensure a clean shutdown process.