Added support for listening for connection changes
The connection listener will be called whenever the socket connection
changes status. This allows the caller to keep track of the connection,
for instance if the Chromecast device is powered off or there are
network issues.

Note: The initial status when setting up the cast device cannot be
caught as the cast object has not yet been returned to the caller.