unicode fix
Explicitly set the encoding of the response to UTF-8, which it (what I understand from the standard) should be anyway.

I don't know about the new Chromecast, but my first generation Chromecast doesn't return any encoding in the HTTP response. Example: {'Content-Length': '1100', 'Application-URL': '<url>', 'Content-Type': 'application/xml'}, and it seems like the fallback to guessing the encoding fails, since the response containing UTF-8 encoded Swedish characters (the assigned name of my Chromecast) gets detected as iso-8859-2

The DIAL specification, 6.1.2 Server Response: "The MIME type of the response SHALL be text/xmland the character encoding SHALL be UTF-8 and SHALL be explicitly indicated with the charsetMIME parameter."

The requests library, Encodings (http://docs.python-requests.org/en/latest/user/advanced/#encodings): "When you receive a response, Requests makes a guess at the encoding to use for decoding the response when you access the Response.text attribute. Requests will first check for an encoding in the HTTP header, and if none is present, will use chardet to attempt to guess the encoding."