Demoted many log levels in socket client.
1) Changed logging for repeated connection failures. First attempt is
logged as EXCEPTION, the rest are logged as DEBUG. If not connection is
made and tries expire, an error is still logged at ERROR.
2) All issues the cause the socket to be reset (read/write errors) are
now logged at INFO and the general message was demoted to WARNING as
the IO errors are handled exceptions that will not generally cause the
application to crash.
3) Unhandled messages to defined namespaces have been demoted to DEBUG.
This was done to match the logging level used to indicate an unhandled
namespace.
4) Status updates have been demoted to DEBUG as they might be too
verbose for INFO.