Ignore benign error in destructor. Fix #113
It’s not really an issue, but when the Chromecast object gets garbage
collected after failing to connect to a device (before creating a
SocketClient instance), its __del__ is called, which in turn calls
`self.socket_client.stop.set()`. This leads to the error described in
the mentioned issue being logged/printed.