Add playQueue, playlist, and list support. All with startItem option.
This adds support for playing additional types of media to Chromecasts with the play methods. Adds list, playQueue, and playlist support (possibly just fixes playlists, they weren't working for me before). This also adds the ability to start at any point in the list/queue/playlist by providing the media item to start on as startItem. This is thanks to and reliant on recent changes to Python-PlexAPI's playQueues by @jjlawren.