Update dial.py
cast type errors from zeroconf, for starters, arent very verbose so printing warnings for them doesnt really add anything except more messages in the client. I would propose relegating these over to the debug channel instead of the warning channel so they dont show up in the console anymore. The errors are handled, and the cast type is overridden anyway and set to CAST_TYPE_CHROMECAST, so it seems kind of silly to warn on these errors. 

At a minimum it would be nice if the error text was included in the warning log so that the user has some type of idea what is going on and how they can potentially fix the problem. I am not even really personally clear on what the problem is I didnt bother to log the actual error, instead I just cut it over to debug channel and let it be.